requirements.txt
numpy>=1.24
pandas>=2.0
PyYAML>=6.0
tqdm>=4.66
matplotlib>=3.7
scikit-learn>=1.3
Pillow>=10.0
opencv-python>=4.8
timm>=0.9
